+++ allauth/account/admin.py
@@ -1,20 +1,25 @@
+import django
 from django.contrib import admin
 
 from .models import EmailConfirmation, EmailAddress
-from . import app_settings
-from ..utils import get_user_model
+from .adapter import get_adapter
 
-User = get_user_model()
 
 class EmailAddressAdmin(admin.ModelAdmin):
     list_display = ('email', 'user', 'primary', 'verified')
     list_filter = ('primary', 'verified')
-    search_fields = ['email'] + list(map(lambda a: 'user__' + a,
-                                    filter(lambda a: a and hasattr(User(), a),
-                                           [app_settings.USER_MODEL_USERNAME_FIELD,
-                                            'first_name',
-                                            'last_name'])))
+    search_fields = []
     raw_id_fields = ('user',)
+
+    def __init__(self, *args, **kwargs):
+        super(EmailAddressAdmin, self).__init__(*args, **kwargs)
+        if not self.search_fields and django.VERSION[:2] < (1, 7):
+            self.search_fields = self.get_search_fields(None)
+
+    def get_search_fields(self, request):
+        base_fields = get_adapter().get_user_search_fields()
+        return ['email'] + list(map(lambda a: 'user__' + a, base_fields))
+
 
 class EmailConfirmationAdmin(admin.ModelAdmin):
     list_display = ('email_address', 'created', 'sent', 'key')