+++ allauth/socialaccount/providers/base.py
@@ -10,11 +10,18 @@ from ..adapter import get_adapter
 class AuthProcess(object):
     LOGIN = 'login'
     CONNECT = 'connect'
+    REDIRECT = 'redirect'
 
 
 class AuthAction(object):
     AUTHENTICATE = 'authenticate'
     REAUTHENTICATE = 'reauthenticate'
+
+
+class AuthError(object):
+    UNKNOWN = 'unknown'
+    CANCELLED = 'cancelled'  # Cancelled on request of user
+    DENIED = 'denied'  # Denied by server
 
 
 class Provider(object):
@@ -51,9 +58,9 @@ class Provider(object):
         email_addresses = self.extract_email_addresses(response)
         self.cleanup_email_addresses(common_fields.get('email'),
                                      email_addresses)
-        sociallogin = SocialLogin(socialaccount,
+        sociallogin = SocialLogin(account=socialaccount,
                                   email_addresses=email_addresses)
-        user = socialaccount.user = adapter.new_user(request, sociallogin)
+        user = sociallogin.user = adapter.new_user(request, sociallogin)
         user.set_unusable_password()
         adapter.populate_user(request, sociallogin, common_fields)
         return sociallogin