+++ docs/changelog.rst
@@ -7,6 +7,11 @@ This chapter contains notes on upgrading.
 From 0.18.0
 ***********
 
+- In the upcoming Django 1.8 it is no longer possible to hookup an
+  unsaved `User` instance to a `SocialAccount`. Therefore, if you are
+  inspecting the `sociallogin` object, you should now use
+  `sociallogin.user` instead of `sociallogin.account.user`.
+
 - When users logged in while `User.is_active` was `False`, they were
   sent to `/accounts/inactive/` in case of a social login, and
   received a form validation error in case of a local login. This